upload big file using firebase storage

+++ functions/src/services/fileSvc.js
@@ -3,21 +3,21 @@
         path = require('path'),
         Busboy = require('busboy'),
         fs = require('fs'),
-        UUID = require("uuid-v4"),
         getRawBody = require('raw-body'),
         contentType = require('content-type'),
         { Storage } = require('@google-cloud/storage'),
-        gcconfig = {
-            projectId: 'm2meme',
-            keyFilename: 'm2meme-firebase-adminsdk-rvmhz-dff76c1bfa.json'
-        },
-        gcs = new Storage(gcconfig),
         ffmpegPath = require('@ffmpeg-installer/ffmpeg').path,
         ffmpeg = require('fluent-ffmpeg'),
-        environment = require('../../env.json')[process.env.NODE_ENV || 'development'];
+        environment = require('../../env.json')[process.env.NODE_ENV || 'development'],
+        MAX_FILE = environment.MAX_FILE,
+        gcconfig = {
+            projectId: environment.projectId,
+            keyFilename: environment.keyFileName
+        },
+        gcs = new Storage(gcconfig);
     //{ exec } = require("child_process");
 
-    var today = new Date(),
+    let today = new Date(),
         filePath = today.getFullYear() + "/" + today.getMonth() + "/" + today.getDate() + "/";
 
     ffmpeg.setFfmpegPath(ffmpegPath);
@@ -37,7 +37,7 @@
                 req,
                 {
                     length: req.headers['content-length'],
-                    limit: '50mb',
+                    limit: MAX_FILE + 'mb',
                     encoding: contentType.parse(req).parameters.charset,
                 },
                 (err, string) => {
@@ -56,11 +56,10 @@
     function middleF2(req, res, next) {
         if (req.method === 'POST' &&
             req.headers['content-type'].startsWith('multipart/form-data')) {
-            let uuid = UUID();
             const busboy = new Busboy({
                 headers: req.headers,
                 limits: {
-                    fileSize: 50 * 1024 * 1024
+                    fileSize: MAX_FILE * 1024 * 1024
                 }
             });
 
@@ -82,8 +81,9 @@
                     };
                     req.file = file_object;
                 });
-                const filepath = path.join(os.tmpdir(), uuid + path.extname(filename));
+                const filepath = path.join(os.tmpdir(), req.user.username + "_" + filename);
                 req.data.file = filepath;
+                req.data.fileName = filename;
                 file.pipe(fs.createWriteStream(filepath));
             });
 
@@ -92,28 +92,29 @@
             });
 
             busboy.on('finish', () => {
-                const bucket = gcs.bucket(environment.FIREBASE_BUCKET),
-                    baseNameOfFile = path.basename(req.data.file);
+                const bucket = gcs.bucket(environment.FIREBASE_BUCKET);
+                let destination = filePath + req.user.username + "/";
 
                 bucket.upload(req.data.file, {
-                    uploadType: 'media',
-                    destination: filePath + baseNameOfFile,
+                    uploadType: 'multipart',
+                    destination: destination + req.data.fileName,
                     resumable: false,
-                    public: true,
                     metadata: { gzip: true, cacheControl: "public, max-age=31536000" }
                 }).then(async uploadedNonMp4File => {
                     if (req.file.mimetype.startsWith('video')) {
-                        await createPoster(bucket, path, baseNameOfFile, req.data.file);
+                        await createPoster(bucket, destination, path, req.user.username, req.data.fileName, req.data.file);
                         fs.unlinkSync(req.data.file);
                     }
-                    return res.status(200).json({
-                        fileLocation: environment.FILE_LOCATION + uploadedNonMp4File[0].name,
-                        filename: uploadedNonMp4File[0].name
-                    });
+                    return res.end();
                 }).catch(err => {
                     console.log("Upload error: " + err);
                     return res.status(500).json(err);
                 });
+
+                return res.write(JSON.stringify({
+                    fileLocation: environment.FILE_LOCATION + destination + req.data.fileName,
+                    filename: req.data.fileName
+                }));
             });
             if (process.env.NODE_ENV) {
                 busboy.end(req.rawBody); // only on production
@@ -125,15 +126,14 @@
         }
     }
 
-    async function createPoster(bucket, path, baseNameOfFile, tempLocalFile) {
-        const gifPosterPath = baseNameOfFile.replace(/\.[^/.]+$/, '_poster.jpg');
-        const posterPath = path.join(os.tmpdir(), gifPosterPath);
+    async function createPoster(bucket, destination, path, username, fileName, tempLocalFile) {
+        const gifPosterPath = fileName.replace(/\.[^/.]+$/, '_poster.jpg');
+        const posterPath = path.join(os.tmpdir(), username + "_" + gifPosterPath);
         await createPosterFromVideo(tempLocalFile, posterPath);
         await bucket.upload(posterPath, {
-            destination: filePath + gifPosterPath,
+            destination: destination + gifPosterPath,
             uploadType: 'media',
             resumable: false,
-            public: true,
             metadata: { gzip: true, cacheControl: "public, max-age=31536000" }
         });
         fs.unlinkSync(posterPath);
@@ -168,24 +168,27 @@
 
     function deleteFile(filename, fileType) {
         // Create a reference to the file to delete
-        const basename = path.basename(filename);
-        const dir = path.dirname(filename);
-        const THUMB_PREFIX = 'thumb_'
-        const bucket = gcs.bucket(environment.FIREBASE_BUCKET);
-        if (fileType.startsWith('image')) {
-            var thumb_file_name = `${THUMB_PREFIX}${basename}`;
-            bucket.file(dir + "/" + thumb_file_name).delete();
-        } else if (fileType.startsWith('video')) {
-            var mp4_file = filename.replace(/\.[^.]+$/, "_output.mp4");
-            bucket.file(mp4_file).delete();
-            var mp4_file_thumb = filename.replace(/\.[^.]+$/, "_thumb_output.mp4");
-            bucket.file(mp4_file_thumb).delete();
-            var poster = filename.replace(/\.[^.]+$/, "_poster.jpg");
-            bucket.file(poster).delete();
-            var posterGif = filename.replace(/\.[^.]+$/, "_poster.gif");
-            bucket.file(posterGif).delete();
+        if (filename) {
+            const basename = path.basename(filename);
+            const dir = path.dirname(filename);
+            const THUMB_PREFIX = 'thumb_'
+            const bucket = gcs.bucket(environment.FIREBASE_BUCKET);
+            if (fileType.startsWith('image')) {
+                var thumb_file_name = `${THUMB_PREFIX}${basename}`;
+                bucket.file(dir + "/" + thumb_file_name).delete();
+            } else if (fileType.startsWith('video')) {
+                var mp4_file = filename.replace(/\.[^.]+$/, "_output.mp4");
+                bucket.file(mp4_file).delete();
+                var mp4_file_thumb = filename.replace(/\.[^.]+$/, "_thumb_output.mp4");
+                bucket.file(mp4_file_thumb).delete();
+                var poster = filename.replace(/\.[^.]+$/, "_poster.jpg");
+                bucket.file(poster).delete();
+                var posterGif = filename.replace(/\.[^.]+$/, "_poster.gif");
+                bucket.file(posterGif).delete();
+            }
+            return bucket.file(filename).delete();
         }
-        return bucket.file(filename).delete();
+        return null;
     }
 
     function deleteByUrl(url, fileType) {
